Update battenberg.R
In the case of a paired analysis, the skip_allele_counting_normal variable is set to TRUE if the number of sampleidx is greater than 1; otherwise, it is set to FALSE.

For scenarios where the user has one tumour and one normal sample pair, the skip_allele_counting_normal variable is always set to TRUE at line 179. However, if the user only has the allele frequency files and not the BAM files, this will cause the script to crash.

To address this issue, I propose the implementation of a new variable, skip_allele_counting_normal. This variable is initially set based on the user-provided skip_allele_counting value or defaults to TRUE if no value is provided. After the first iteration of nsamples, the skip_allele_counting_normal variable is updated to FALSE to prevent the recreation of the normal allele frequency files in subsequent iterations.